refactor: extract module-level SCL helpers into src/utils/ied-scl.ts
